refactor: lift the field-run machinery into shared decorator and member modules
- Adopt `is_const_true` and `StmtClassDef::bases()`, retiring the hand-unwrapped boolean literal and base-list walks
- Lift `crate/src/primitives/decorator.rs` holding both readers, so `params.rs` and `constructor.rs` share one decorator-arguments read
- Extract `crate/src/rules/alphabetize/members.rs` for the five class-body classifiers `mod.rs` carried beside its rewrite recursion
- Narrow `decorator_arguments` to `pub(super)`, both consumers sitting inside `primitives/`
- Add fixtures pinning method sorting inside a pinned class and a `kw_only` base's field sort
- Split Pydantic across the rule page and two glossary entries, `pydantic.dataclasses` holding whereas `BaseModel` sorts
- Give the landing demo's `@dataclass` a `kw_only=True`, so its field sort still demonstrates
