+++ crate/src/rules/alphabetize/members.rs
@@ -1,0 +1,127 @@
+//! Class-body member classifiers. Reads a statement as an annotated
+//! field or a simple assignment, reports whether a field carries a
+//! default, and reports the method group and the decorator condition
+//! that pins a method run.
+
+use ruff_python_ast::{
+    Stmt, StmtAnnAssign, StmtFunctionDef,
+    helpers::{any_over_expr, is_dunder},
+};
+
+use crate::primitives::{
+    binding::{annotated_name_target, single_name_target},
+    decorator::decorator_simple_name,
+    params::pins_positional_params,
+};
+
+/// Returns the `StmtAnnAssign` and its target name when the target
+/// is a single `Name`.
+pub(super) fn ann_assign_with_named_field(stmt: &Stmt) -> Option<(&StmtAnnAssign, &str)> {
+    let ann = stmt.as_ann_assign_stmt()?;
+    Some((ann, annotated_name_target(ann)?))
+}
+
+/// True when a class body has at least two `Stmt::AnnAssign` field
+/// declarations and at least one method whose decorator carries
+/// positional arguments.
+pub(super) fn class_pins_methods(body: &[Stmt]) -> bool {
+    body.iter()
+        .filter(|s| ann_assign_with_named_field(s).is_some())
+        .nth(1)
+        .is_some()
+        && body
+            .iter()
+            .filter_map(Stmt::as_function_def_stmt)
+            .any(pins_positional_params)
+}
+
+/// True when an annotated assignment carries a default, either
+/// directly via `= value` or through any nested `Call` in the
+/// annotation that carries a `default` or `default_factory` keyword.
+pub(super) fn has_default(ann: &StmtAnnAssign) -> bool {
+    ann.value.is_some()
+        || any_over_expr(&ann.annotation, |e| {
+            e.as_call_expr().is_some_and(|c| {
+                c.arguments
+                    .keywords
+                    .iter()
+                    .any(|kw| matches!(kw.arg.as_deref(), Some("default" | "default_factory")))
+            })
+        })
+}
+
+/// Returns the method-group index. `0` for dunders, `1` for
+/// `@property` / `@cached_property` (decided by the first decorator),
+/// `2` for single-leading-underscore privates, `3` for public.
+pub(super) fn method_group(f: &StmtFunctionDef) -> u8 {
+    let name = f.name.as_str();
+    if is_dunder(name) {
+        0
+    } else if f
+        .decorator_list
+        .first()
+        .and_then(decorator_simple_name)
+        .is_some_and(|n| matches!(n, "cached_property" | "property"))
+    {
+        1
+    } else if name.starts_with('_') {
+        2
+    } else {
+        3
+    }
+}
+
+/// Returns the simple name assigned by an `Stmt::Assign` whose
+/// target is a single `Name`. `None` for multi-target,
+/// destructuring, attribute, or subscript targets.
+pub(super) fn simple_name_assign(stmt: &Stmt) -> Option<&str> {
+    single_name_target(stmt.as_assign_stmt()?)
+}
+
+#[cfg(test)]
+mod tests {
+    use indoc::indoc;
+
+    use super::*;
+    use crate::testing::{first_class, parse};
+
+    #[test]
+    fn ann_assign_with_named_field_filters_to_name_targets() {
+        let s = parse("x: int = 1\nself.x: int = 1\n");
+        let names: Vec<Option<&str>> = s
+            .ast()
+            .body
+            .iter()
+            .map(|s| ann_assign_with_named_field(s).map(|(_, name)| name))
+            .collect();
+        assert_eq!(names, vec![Some("x"), None]);
+    }
+
+    #[test]
+    fn method_group_orders_dunder_property_private_public() {
+        let src = indoc! {"
+            class C:
+                def __init__(self): pass
+                @property
+                def name(self): pass
+                def _helper(self): pass
+                def public(self): pass
+        "};
+        let s = parse(src);
+        let class = first_class(&s);
+        let groups: Vec<u8> = class
+            .body
+            .iter()
+            .filter_map(Stmt::as_function_def_stmt)
+            .map(method_group)
+            .collect();
+        assert_eq!(groups, vec![0, 1, 2, 3]);
+    }
+
+    #[test]
+    fn simple_name_assign_filters_to_single_name_targets() {
+        let s = parse("X = 1\nself.x = 1\nx, y = 1, 2\n");
+        let names: Vec<Option<&str>> = s.ast().body.iter().map(simple_name_assign).collect();
+        assert_eq!(names, vec![Some("X"), None, None]);
+    }
+}
