+++ crate/src/rules/alphabetize/mod.rs
@@ -23,16 +23,12 @@
 use std::borrow::Cow;
 
 use ruff_diagnostics::Edit;
-use ruff_python_ast::{
-    Stmt, StmtAnnAssign, StmtFunctionDef,
-    helpers::{any_over_expr, is_compound_statement, is_dunder},
-};
+use ruff_python_ast::{Stmt, helpers::is_compound_statement};
 use ruff_text_size::{Ranged, TextRange, TextSize};
 
 use crate::{
     config::Config,
     primitives::{
-        binding::{annotated_name_target, decorator_simple_name, single_name_target},
         constructor::keyword_field_start,
         edit::{apply_inline_edits, singleton_groups, splice_bodies},
         imports::{defers_annotations, import_blank_lines, import_sort_key, sectioned_import_runs},
@@ -40,7 +36,6 @@ use crate::{
             adjacent_slots, any_sibling_shares_line, assemble_or_borrow, assembled_cell_edits,
             permute_runs, rendered_member_blocks,
         },
-        params::pins_positional_params,
         scope::{BodyScope, compound_sub_bodies, scoped_body},
         sections::Sections,
         tiering::permute_defs,
@@ -52,10 +47,12 @@ use crate::{
 mod class_graph;
 mod dict;
 mod leaves;
+mod members;
 
 use self::{
     class_graph::permute_class_assigns,
     leaves::{collect_docstring_entry_edits, collect_leaf_edits},
+    members::{class_pins_methods, method_group},
 };
 
 pub(crate) struct Alphabetize {
@@ -145,13 +142,6 @@ struct RewriteCtx<'a> {
     source: &'a Source,
 }
 
-/// Returns the `StmtAnnAssign` and its target name when the target
-/// is a single `Name`.
-fn ann_assign_with_named_field(stmt: &Stmt) -> Option<(&StmtAnnAssign, &str)> {
-    let ann = stmt.as_ann_assign_stmt()?;
-    Some((ann, annotated_name_target(ann)?))
-}
-
 /// Computes the reorder of `body`: renders each member, then permutes the
 /// slots within each section by the family sorts and import grouping that
 /// `scope` enables, leaving the assembly to the caller. The section
@@ -232,60 +222,10 @@ fn body_layout<'a>(
     }
 }
 
-/// True when a class body has at least two `Stmt::AnnAssign` field
-/// declarations and at least one method whose decorator carries
-/// positional arguments.
-fn class_pins_methods(body: &[Stmt]) -> bool {
-    body.iter()
-        .filter(|s| ann_assign_with_named_field(s).is_some())
-        .nth(1)
-        .is_some()
-        && body
-            .iter()
-            .filter_map(Stmt::as_function_def_stmt)
-            .any(pins_positional_params)
-}
-
-/// True when an annotated assignment carries a default, either
-/// directly via `= value` or through any nested `Call` in the
-/// annotation that carries a `default` or `default_factory` keyword.
-fn has_default(ann: &StmtAnnAssign) -> bool {
-    ann.value.is_some()
-        || any_over_expr(&ann.annotation, |e| {
-            e.as_call_expr().is_some_and(|c| {
-                c.arguments
-                    .keywords
-                    .iter()
-                    .any(|kw| matches!(kw.arg.as_deref(), Some("default" | "default_factory")))
-            })
-        })
-}
-
 /// The one-newline divider an import-run collapse inserts after new-order
 /// slot `i`, `None` where the neighbors do not collapse onto one line.
 fn import_gap(import_run_slots: &[usize], i: usize) -> Option<&'static str> {
     import_run_slots.binary_search(&i).is_ok().then_some("\n")
-}
-
-/// Returns the method-group index. `0` for dunders, `1` for
-/// `@property` / `@cached_property` (decided by the first decorator),
-/// `2` for single-leading-underscore privates, `3` for public.
-fn method_group(f: &StmtFunctionDef) -> u8 {
-    let name = f.name.as_str();
-    if is_dunder(name) {
-        0
-    } else if f
-        .decorator_list
-        .first()
-        .and_then(decorator_simple_name)
-        .is_some_and(|n| matches!(n, "cached_property" | "property"))
-    {
-        1
-    } else if name.starts_with('_') {
-        2
-    } else {
-        3
-    }
 }
 
 /// Rewrites a non-empty body, returning the rewritten text alongside
@@ -356,31 +296,12 @@ fn rewrite_stmt<'a>(
     splice_bodies(ctx.source, block, [(body_text, body_span)], ctx.leaf_edits)
 }
 
-/// Returns the simple name assigned by an `Stmt::Assign` whose
-/// target is a single `Name`. `None` for multi-target,
-/// destructuring, attribute, or subscript targets.
-fn simple_name_assign(stmt: &Stmt) -> Option<&str> {
-    single_name_target(stmt.as_assign_stmt()?)
-}
-
 #[cfg(test)]
 mod tests {
     use indoc::indoc;
 
     use super::*;
-    use crate::testing::{applied_text, first_class, parse};
-
-    #[test]
-    fn ann_assign_with_named_field_filters_to_name_targets() {
-        let s = parse("x: int = 1\nself.x: int = 1\n");
-        let names: Vec<Option<&str>> = s
-            .ast()
-            .body
-            .iter()
-            .map(|s| ann_assign_with_named_field(s).map(|(_, name)| name))
-            .collect();
-        assert_eq!(names, vec![Some("x"), None]);
-    }
+    use crate::testing::{applied_text, parse};
 
     #[test]
     fn apply_skips_docstring_entry_reorder_when_config_disables_it() {
@@ -410,33 +331,5 @@ mod tests {
             bar_pos < alpha_pos,
             "docstring entries should keep source order when sort-docstring-entries is off",
         );
-    }
-
-    #[test]
-    fn method_group_orders_dunder_property_private_public() {
-        let src = indoc! {"
-            class C:
-                def __init__(self): pass
-                @property
-                def name(self): pass
-                def _helper(self): pass
-                def public(self): pass
-        "};
-        let s = parse(src);
-        let class = first_class(&s);
-        let groups: Vec<u8> = class
-            .body
-            .iter()
-            .filter_map(Stmt::as_function_def_stmt)
-            .map(method_group)
-            .collect();
-        assert_eq!(groups, vec![0, 1, 2, 3]);
-    }
-
-    #[test]
-    fn simple_name_assign_filters_to_single_name_targets() {
-        let s = parse("X = 1\nself.x = 1\nx, y = 1, 2\n");
-        let names: Vec<Option<&str>> = s.ast().body.iter().map(simple_name_assign).collect();
-        assert_eq!(names, vec![Some("X"), None, None]);
     }
 }
