+++ crate/tests/fixtures/alphabetize/msgspec_struct_kw_only_fields_sort/input.py
@@ -1,0 +1,7 @@
+import msgspec
+
+
+class Posting(msgspec.Struct, kw_only=True):
+    title: str
+    company: str
+    date_posted: str
