+++ crate/tests/fixtures/alphabetize/pinned_class_methods_still_sort/input.py
@@ -1,0 +1,13 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class Posting:
+    title: str
+    company: str
+
+    def summary(self):
+        return self.title
+
+    def render(self):
+        return self.company
